hook personalizado
